create repositories to insert many

+++ app/models/casualty_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from app.models import Base
 
@@ -7,8 +7,8 @@ class Casualty(Base):
     __tablename__ = 'casualties'
 
     id = Column(Integer, primary_key=True, autoincrement=True, nullable=False)
-    killed = Column(Integer)
-    wounded = Column(Integer)
+    killed = Column(Float)
+    wounded = Column(Float)
 
     attack_id = Column(Integer, ForeignKey("attacks.id"))
     attack = relationship("Attack", back_populates="casualties")
